requirements.txt
amqp==5.0.9
appnope==0.1.2
APScheduler==3.6.3
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
billiard==3.6.4.0
black==22.1.0
cachetools==4.2.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-timezone-field==4.2.3
executing==0.8.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
ipython==8.0.1
jedi==0.18.1
kombu==5.2.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
